Remove Multiple Inheritance
Besides being an anitpattern this is interfering with updateing the python
bindings to pybind11.

- move getKneeReweightingModel() to FluxFunction with a default to throw
  if model is not "none". Reweightable fluxes override this
- move pion and kaon relative contributions are moved to LegacyConventional
  as those are the only models which use them
