+++ src/include/nuflux/meson.build
@@ -1,6 +1,6 @@
 nuflux_headers = [
   'ANFlux.h', 'Fluxes.h', 'FluxFunction.h',
-  'Interfaces.h', 'IPLEFlux.h', 'LegacyConventionalFlux.h',
+  'IPLEFlux.h', 'LegacyConventionalFlux.h',
   'LegacyPromptFlux.h', 'nuflux.h',
   'SplineFlux2.h', 'SplineFlux.h',
 ]
